Update migration files

+++ db/migrate/20231120132353_create_recipes.rb
@@ -3,8 +3,8 @@ class CreateRecipes < ActiveRecord::Migration[7.1]
     create_table :recipes do |t|
       t.references :user, null: false, foreign_key: { on_delete: :cascade, on_update: :cascade}
       t.string :name
-      t.decimal :preparation_time
-      t.decimal :cooking_time
+      t.integer :preparation_time
+      t.integer :cooking_time
       t.text :description
       t.boolean :public
